firmware/test/integration/requirements.txt
websocket-client>=1.6.1
